clean our code
