--- python-file/games.py
@@ -1,12 +1,0 @@
-from botato import *
-import spacy
-
-
-
-
-if __name__ == '__main__':
-    #botato functions
-    root = Little()
-    root.title('Little Kiddie')
-    root.mainloop()
-    #################
